requirements.txt
paramiko==2.3.1
tornado==4.5.2
PyYAML==3.12
path.py==10.4
git+git://github.com/a598824322/pyping.git#egg=pyping
aelog
